clean-up the new validation rules structure
