+++ packages/graphql/src/schema/validation/custom-rules/directives/default.ts
@@ -20,13 +20,13 @@
 import type { ASTVisitor, EnumTypeDefinitionNode, FieldDefinitionNode, TypeNode } from "graphql";
 import { Kind } from "graphql";
 import { GraphQLDate } from "graphql-compose";
-import { GRAPHQL_BUILTIN_SCALAR_TYPES } from "../../../../../constants";
-import { defaultDirective } from "../../../../../graphql/directives";
-import { GraphQLDateTime, GraphQLLocalDateTime, GraphQLLocalTime, GraphQLTime } from "../../../../../graphql/scalars";
-import type { Neo4jValidationContext, ObjectExtensionsTypeMap } from "../../../Neo4jValidationContext";
-import { assertValid, createGraphQLError, DocumentValidationError } from "../../utils/document-validation-error";
-import { getPathToNode } from "../../utils/path-parser";
-import { assertArgumentHasSameTypeAsField } from "../../utils/same-type-argument-as-field";
+import { GRAPHQL_BUILTIN_SCALAR_TYPES } from "../../../../constants";
+import { defaultDirective } from "../../../../graphql/directives";
+import { GraphQLDateTime, GraphQLLocalDateTime, GraphQLLocalTime, GraphQLTime } from "../../../../graphql/scalars";
+import type { Neo4jValidationContext, ObjectExtensionsTypeMap } from "../../Neo4jValidationContext";
+import { assertValid, createGraphQLError, DocumentValidationError } from "../utils/document-validation-error";
+import { getPathToNode } from "../utils/path-parser";
+import { assertArgumentHasSameTypeAsField } from "../utils/same-type-argument-as-field";
 import { fieldIsInNodeType, fieldIsInRelationshipPropertiesType } from "./check-if-location-is-valid";
 
 export function validateDefaultDirective(context: Neo4jValidationContext): ASTVisitor {
@@ -34,7 +34,7 @@ export function validateDefaultDirective(context: Neo4jValidationContext): ASTVi
     if (!extensionsTypeMap) {
         throw new Error("No extensionsTypeMap found in the context");
     }
-    // TODO: if this is needed to other validation rules, we could move these to the context
+
     const enumsTypes: EnumTypeDefinitionNode[] = Object.values(extensionsTypeMap)
         .map((type) => type.definition)
         .filter((definition): definition is EnumTypeDefinitionNode => definition.kind === Kind.ENUM_TYPE_DEFINITION);
