--- packages/graphql/src/schema/validation/custom-rules/directives/test-rules/limit.ts
@@ -1,254 +1,0 @@
-/*
- * Copyright (c) "Neo4j"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { ASTVisitor, InterfaceTypeDefinitionNode, ObjectTypeDefinitionNode } from "graphql";
-import { limitDirective } from "../../../../../graphql/directives";
-import { asArray } from "../../../../../utils/utils";
-import type { Neo4jValidationContext } from "../../../Neo4jValidationContext";
-import { assertValid, createGraphQLError, DocumentValidationError } from "../../utils/document-validation-error";
-import { parseArgumentToInt } from "../../utils/utils";
-import { typeIsANodeType } from "./check-if-location-is-valid";
-
-// function validateFulltextDirective(context: Neo4jValidationContext): ASTVisitor {
-//     const extensionsTypeMap = context.extensionsTypeMap;
-//     if (!extensionsTypeMap) {
-//         throw new Error("No extensionsTypeMap found in the context");
-//     }
-//     return {
-//         ObjectTypeDefinition(objectTypeDefinitionNode: ObjectTypeDefinitionNode, _key, _parent, _path, _ancestors) {
-//             const { directives } = objectTypeDefinitionNode;
-//             const objectTypeExtensionNodes = extensionsTypeMap[objectTypeDefinitionNode.name.value]?.extensions;
-//             const extensionsDirectives = asArray(objectTypeExtensionNodes).flatMap((extensionNode) => {
-//                 return extensionNode.directives ?? [];
-//             });
-//             const allDirectives = [...(directives ?? []), ...extensionsDirectives];
-//             const appliedFulltextDirective = allDirectives.find(
-//                 (directive) => directive.name.value === fulltextDirective.name
-//             );
-//             if (!appliedFulltextDirective) {
-//                 return;
-//             }
-//             const indexesArg = appliedFulltextDirective.arguments?.find((a) => a.name.value === "indexes");
-//             if (!indexesArg) {
-//                 // delegate to DirectiveArgumentOfCorrectType rule
-//                 return;
-//             }
-
-//             const compatibleFields = getFulltextCompatibleFields(objectTypeDefinitionNode);
-//             const isValidLocation = typeIsANodeType({ objectTypeDefinitionNode, extensionsTypeMap });
-//             const { isValid, errorMsg, errorPath } = assertValid(() => {
-//                 if (!isValidLocation) {
-//                     throw new DocumentValidationError(
-//                         `Directive "${fulltextDirective.name}" requires to be used within the "@node" directive`,
-//                         []
-//                     );
-//                 }
-//                 const indexesValues = parseValueNode(indexesArg.value) as FulltextField[];
-
-//                 indexesValues.forEach((indexValue) => {
-//                     const indexName = indexValue.indexName;
-//                     const indexNames = indexesValues.filter((i) => indexName === i.indexName);
-//                     if (indexNames.length > 1) {
-//                         throw new DocumentValidationError(
-//                             `@fulltext.indexes invalid value for: ${indexName}. Duplicate index name.`,
-//                             ["indexes"]
-//                         );
-//                     }
-
-//                     const queryName = indexValue.queryName;
-//                     const queryNames = indexesValues.filter((i) => queryName === i.queryName);
-//                     if (queryNames.length > 1) {
-//                         throw new DocumentValidationError(
-//                             `@fulltext.indexes invalid value for: ${queryName}. Duplicate query name.`,
-//                             ["indexes"]
-//                         );
-//                     }
-//                     asArray(indexValue.fields).forEach((indexField) => {
-//                         if (!compatibleFields[indexField]) {
-//                             throw new DocumentValidationError(
-//                                 `@fulltext.indexes invalid value for: ${indexValue.indexName}. Field ${indexField} is not of type String or ID.`,
-//                                 ["indexes"]
-//                             );
-//                         }
-//                     });
-//                 });
-//             });
-//             if (!isValid) {
-//                 context.reportError(
-//                     createGraphQLError({
-//                         nodes: [objectTypeDefinitionNode],
-//                         path: [objectTypeDefinitionNode.name.value, `@${fulltextDirective.name}`, ...errorPath],
-//                         errorMsg,
-//                     })
-//                 );
-//             }
-//         },
-//     };
-// }
-
-export function validateLimitDirective(context: Neo4jValidationContext): ASTVisitor {
-    const extensionsTypeMap = context.extensionsTypeMap;
-    if (!extensionsTypeMap) {
-        throw new Error("No extensionsTypeMap found in the context");
-    }
-    return {
-        InterfaceTypeDefinition(
-            interfaceTypeDefinitionNode: InterfaceTypeDefinitionNode,
-            _key,
-            _parent,
-            _path,
-            _ancestors
-        ) {
-            const { directives } = interfaceTypeDefinitionNode;
-            const objectTypeExtensionNodes = extensionsTypeMap[interfaceTypeDefinitionNode.name.value]?.extensions;
-            const extensionsDirectives = asArray(objectTypeExtensionNodes).flatMap((extensionNode) => {
-                return extensionNode.directives ?? [];
-            });
-            const allDirectives = [...(directives ?? []), ...extensionsDirectives];
-            const appliedLimitDirective = allDirectives.find(
-                (directive) => directive.name.value === limitDirective.name
-            );
-            if (!appliedLimitDirective) {
-                return;
-            }
-            const defaultArg = appliedLimitDirective.arguments?.find((a) => a.name.value === "default");
-            const maxArg = appliedLimitDirective.arguments?.find((a) => a.name.value === "max");
-            if (!defaultArg && !maxArg) {
-                // nothing to check, fields are optional
-                return;
-            }
-            const defaultLimit = parseArgumentToInt(defaultArg);
-            const maxLimit = parseArgumentToInt(maxArg);
-            const { isValid, errorMsg, errorPath } = assertValid(() => {
-                // TODO: we should check that all the implementing types are of type node
-                if (defaultLimit) {
-                    const defaultValue = defaultLimit.toNumber();
-                    // default must be greater than 0
-                    if (defaultValue <= 0) {
-                        throw new DocumentValidationError(
-                            `@limit.default invalid value: ${defaultValue}. Must be greater than 0.`,
-                            ["default"]
-                        );
-                    }
-                }
-                if (maxLimit) {
-                    const maxValue = maxLimit.toNumber();
-                    // max must be greater than 0
-                    if (maxValue <= 0) {
-                        throw new DocumentValidationError(
-                            `@limit.max invalid value: ${maxValue}. Must be greater than 0.`,
-                            ["max"]
-                        );
-                    }
-                }
-                if (defaultLimit && maxLimit) {
-                    const defaultValue = defaultLimit.toNumber();
-                    const maxValue = maxLimit.toNumber();
-                    // default must be smaller than max
-                    if (maxLimit < defaultLimit) {
-                        throw new DocumentValidationError(
-                            `@limit.max invalid value: ${maxValue}. Must be greater than limit.default: ${defaultValue}.`,
-                            ["max"]
-                        );
-                    }
-                }
-            });
-            if (!isValid) {
-                context.reportError(
-                    createGraphQLError({
-                        nodes: [interfaceTypeDefinitionNode],
-                        path: [interfaceTypeDefinitionNode.name.value, `@${limitDirective.name}`, ...errorPath],
-                        errorMsg,
-                    })
-                );
-            }
-        },
-        ObjectTypeDefinition(objectTypeDefinitionNode: ObjectTypeDefinitionNode, _key, _parent, _path, _ancestors) {
-            const { directives } = objectTypeDefinitionNode;
-            const objectTypeExtensionNodes = extensionsTypeMap[objectTypeDefinitionNode.name.value]?.extensions;
-            const extensionsDirectives = asArray(objectTypeExtensionNodes).flatMap((extensionNode) => {
-                return extensionNode.directives ?? [];
-            });
-            const allDirectives = [...(directives ?? []), ...extensionsDirectives];
-            const appliedLimitDirective = allDirectives.find(
-                (directive) => directive.name.value === limitDirective.name
-            );
-            if (!appliedLimitDirective) {
-                return;
-            }
-            const defaultArg = appliedLimitDirective.arguments?.find((a) => a.name.value === "default");
-            const maxArg = appliedLimitDirective.arguments?.find((a) => a.name.value === "max");
-            if (!defaultArg && !maxArg) {
-                // nothing to check, fields are optional
-                return;
-            }
-            const defaultLimit = parseArgumentToInt(defaultArg);
-            const maxLimit = parseArgumentToInt(maxArg);
-            const isValidLocation = typeIsANodeType({ objectTypeDefinitionNode, extensionsTypeMap });
-            const { isValid, errorMsg, errorPath } = assertValid(() => {
-                if (!isValidLocation) {
-                    throw new DocumentValidationError(
-                        `Directive "${limitDirective.name}" requires to be used within the "@node" directive or in an interface type`,
-                        []
-                    );
-                }
-                if (defaultLimit) {
-                    const defaultValue = defaultLimit.toNumber();
-                    // default must be greater than 0
-                    if (defaultValue <= 0) {
-                        throw new DocumentValidationError(
-                            `@limit.default invalid value: ${defaultValue}. Must be greater than 0.`,
-                            ["default"]
-                        );
-                    }
-                }
-                if (maxLimit) {
-                    const maxValue = maxLimit.toNumber();
-                    // max must be greater than 0
-                    if (maxValue <= 0) {
-                        throw new DocumentValidationError(
-                            `@limit.max invalid value: ${maxValue}. Must be greater than 0.`,
-                            ["max"]
-                        );
-                    }
-                }
-                if (defaultLimit && maxLimit) {
-                    const defaultValue = defaultLimit.toNumber();
-                    const maxValue = maxLimit.toNumber();
-                    // default must be smaller than max
-                    if (maxLimit < defaultLimit) {
-                        throw new DocumentValidationError(
-                            `@limit.max invalid value: ${maxValue}. Must be greater than limit.default: ${defaultValue}.`,
-                            ["max"]
-                        );
-                    }
-                }
-            });
-            if (!isValid) {
-                context.reportError(
-                    createGraphQLError({
-                        nodes: [objectTypeDefinitionNode],
-                        path: [objectTypeDefinitionNode.name.value, `@${limitDirective.name}`, ...errorPath],
-                        errorMsg,
-                    })
-                );
-            }
-        },
-    };
-}
