+++ packages/graphql/src/schema/validation/custom-rules/features/valid-relationship-properties.ts
@@ -92,21 +92,21 @@ function assertRelationshipProperties(traversedDef: ObjectOrInterfaceWithExtensi
             }
         });
 
-        if (field.directives) {
-            const forbiddenDirectives = [
-              //  "authorization",
-               // "authentication",
-                "subscriptionsAuthorization",
-               // "relationship",
-             //  "cypher",
-            ];
-            const foundForbiddenDirective = field.directives.find((d) => forbiddenDirectives.includes(d.name.value));
-            if (foundForbiddenDirective) {
-                throw new DocumentValidationError(
-                    `Invalid @relationshipProperties field: Cannot use the @${foundForbiddenDirective.name.value} directive on relationship properties.`,
-                    errorPath
-                );
-            }
-        }
+        // if (field.directives) {
+        //     const forbiddenDirectives = [
+        //       //  "authorization",
+        //        // "authentication",
+        //        // "subscriptionsAuthorization",
+        //        // "relationship",
+        //      //  "cypher",
+        //     ];
+        //     const foundForbiddenDirective = field.directives.find((d) => forbiddenDirectives.includes(d.name.value));
+        //     if (foundForbiddenDirective) {
+        //         throw new DocumentValidationError(
+        //             `Invalid @relationshipProperties field: Cannot use the @${foundForbiddenDirective.name.value} directive on relationship properties.`,
+        //             errorPath
+        //         );
+        //     }
+        // }
     });
 }
